logger fix

+++ include/logger.hpp
@@ -11,7 +11,7 @@ class Logger
 public:
 
 	static void LogCpuRegisters(const std::unique_ptr<Registers>& registers);
-	static void LogMemory(const std::unique_ptr<Mmu>& memory, uint16_t start, uint16_t size, const char* tag = nullptr);
+	static void LogMemory(const std::unique_ptr<Mmu>& memory, size_t start, size_t size, const char* tag = nullptr);
 	static void Enable() { is_enabled_ = 1; };
 	static void Disable() { is_enabled_ = 0; };
 	static void DebugHelp();
